Changed class "Classes" to Other and moved Transaction class to Other. Removed AccountActivity class (not used)
