+++ RebelAllianceBank/Accounts/SavingsAccount.cs
@@ -1,5 +1,6 @@
-﻿using RebelAllianceBank.Classes;
-using RebelAllianceBank.Interfaces;
+﻿using RebelAllianceBank.Interfaces;
+using RebelAllianceBank.Other;
+
 namespace RebelAllianceBank.Accounts
 {
     public class SavingsAccount : IBankAccount
