+++ RebelAllianceBank/Menu/Menu.cs
@@ -1,5 +1,4 @@
 ﻿using RebelAllianceBank.Interfaces;
-using RebelAllianceBank.utils;
 
 namespace RebelAllianceBank.Menu
 {
