+++ RebelAllianceBank/utils/ExchangeRate.cs
@@ -1,10 +1,5 @@
-using System.Collections.Concurrent;
-using System.Diagnostics;
-using System.Reflection.Metadata.Ecma335;
-using System.Security.Cryptography;
-using System.Threading.Channels;
-using RebelAllianceBank.Classes;
-using RebelAllianceBank.utils;
+using RebelAllianceBank.Other;
+
 
 namespace RebelAllianceBank.utils;
 public class ExchangeRate
