+++ RebelAllianceBank/utils/InlaneColor.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace RebelAllianceBank.utils;
 
 /// <summary>
