+++ RebelAllianceBank/utils/Markdown.cs
@@ -1,6 +1,5 @@
 using System.Text.RegularExpressions;
 using RebelAllianceBank.Enums;
-using System.Text.RegularExpressions;
 
 namespace RebelAllianceBank.utils;
